refactor(split-rendering): quarantine overlays + render_view_lines (Phase 3a-c)
Create orchestration/ subdirectory to physically separate code that shares
the SelectionContext / DecorationContext "mega structs" from the rest of
split_rendering:

- orchestration/contexts.rs: SelectionContext, DecorationContext (data only)
- orchestration/overlays.rs: selection_context, decoration_context producers
- orchestration/render_line.rs: render_view_lines (~1,245-line hot path) +
  LineRenderInput, LineRenderOutput, LastLineEnd

Every top-level module at split_rendering/*.rs still has zero dependency
on these shared carriers; only files inside orchestration/ do.

All 37 split_rendering tests pass.
